Handle vector boolean types when calculating the SLP unroll factor
The SLP unrolling factor is calculated by finding the smallest
scalar type for each SLP statement and taking the number of required
lanes from the vector versions of those scalar types.  E.g. for an
int32->int64 conversion, it's the vector of int32s rather than the
vector of int64s that determines the unroll factor.

However, we made no special allowance for vector boolean types, and
always recalculated the vector type from the narrowest scalar type.
This caused the unroll factor to be overly conservative if vector
boolean types are smaller than vector data types, since we'd use a
scalar type from the former to create an instance of the latter.